Move methods to AbstractFileTransfer
